Fix sign warnings for mulShift_mod1e9().
mulShift_mod1e9() takes `const int32_t j`. This is within [128, 180],
so when we have uint32_t, we can use a value-preserving cast to int32_t.

+++ ryu/d2fixed.c
@@ -402,7 +402,7 @@ int d2fixed_buffered_n(double d, uint32_t precision, char* result) {
       const uint32_t j = p10bits - e2;
       // Temporary: j is usually around 128, and by shifting a bit, we push it to 128 or above, which is
       // a slightly faster code path in mulShift_mod1e9. Instead, we can just increase the multipliers.
-      const uint32_t digits = mulShift_mod1e9(m2 << 8, POW10_SPLIT[POW10_OFFSET[idx] + i], j + 8);
+      const uint32_t digits = mulShift_mod1e9(m2 << 8, POW10_SPLIT[POW10_OFFSET[idx] + i], (int32_t) (j + 8));
       if (nonzero) {
         append_nine_digits(digits, result + index);
         index += 9;
@@ -616,7 +616,7 @@ int d2exp_buffered_n(double d, uint32_t precision, char* result) {
       const uint32_t j = p10bits - e2;
       // Temporary: j is usually around 128, and by shifting a bit, we push it to 128 or above, which is
       // a slightly faster code path in mulShift_mod1e9. Instead, we can just increase the multipliers.
-      digits = mulShift_mod1e9(m2 << 8, POW10_SPLIT[POW10_OFFSET[idx] + i], j + 8);
+      digits = mulShift_mod1e9(m2 << 8, POW10_SPLIT[POW10_OFFSET[idx] + i], (int32_t) (j + 8));
       if (printedDigits != 0) {
         if (printedDigits + 9 > precision) {
           availableDigits = 9;
